Move ASG configs to autoscaling_configuration.go
